+++ snflwrpy/__init__.py
@@ -1,0 +1,4 @@
+"""Module for snflwrpy"""
+__app_name__ = 'snflwrpy'
+__app_version__ = '1.0.0'
+__app_description__ = "Visualize sorting algorithms with Vogel's Fibonacci model."