+++ tests/sorters/test_bubble_sorter.py
@@ -1,22 +1,8 @@
 """
 Test suite for the bubble_sorter module.
 """
-from vogel_sorter.sorters.bubble_sorter import BubbleSorter
-from vogel_sorter.sorters.sort_change import SortChange
+from snflwrpy.sorters.bubble_sorter import BubbleSorter
 from chancepy import Chance
-
-def test_bubble_sorter_next_returns_expected_sort_change():
-    """Test bubble sorter next returns expected sort changes"""
-
-    unsorted_array = [Chance.integer(21, 40), Chance.integer(1,20)]
-    sorter = BubbleSorter(unsorted_array)
-    first_expected_sort_change = SortChange(index=0, old_value=unsorted_array[0], new_value=unsorted_array[1])
-    second_expected_sort_change = SortChange(index=1, old_value=unsorted_array[1], new_value=unsorted_array[0])
-
-    next_results = sorter.next()
-
-    assert next_results[0] == first_expected_sort_change
-    assert next_results[1] == second_expected_sort_change
 
 def test_bubble_sorter_is_sorted_after_iterating_next_n_squared_times():
     """Test bubble sorter is sorted after iterating next n squared times"""