+++ tests/sorters/test_radix_lsd_sorter.py
@@ -1,6 +1,5 @@
 """Test suite for radix_lsd_sorter module"""
-from vogel_sorter.sorters.radix_lsd_sorter import RadixLSDSorter
-from vogel_sorter.sorters.sort_change import SortChange
+from snflwrpy.sorters.radix_lsd_sorter import RadixLSDSorter
 from chancepy import Chance
 
 def test_radix_lsd_sorter_is_sorted_after_iterating_list_length_times_digit_length():