+++ tests/test_sort_type.py
@@ -1,12 +1,13 @@
 """
 Test suite for sort_type module.
 """
-from vogel_sorter.sort_type import SortType
+from snflwrpy.sort_type import SortType
 
 def test_sort_type_returns_list_of_all_values_when_called():
     """Test sort type returns list of all values when called"""
 
     expected_list = [
+        SortType.BOGO.value,
         SortType.BUBBLE.value,
         SortType.COCKTAIL.value,
         SortType.GNOME.value,