+++ tests/test_sorter_factory.py
@@ -1,13 +1,21 @@
 """
 Test suite for sorter_factory module.
 """
-from vogel_sorter.sorter_factory import sorter_factory
-from vogel_sorter.sorters.bubble_sorter import BubbleSorter
-from vogel_sorter.sorters.cocktail_sorter import CocktailSorter
-from vogel_sorter.sorters.gnome_sorter import GnomeSorter
-from vogel_sorter.sorters.insertion_sorter import InsertionSorter
-from vogel_sorter.sorters.radix_lsd_sorter import RadixLSDSorter
-from vogel_sorter.sort_type import SortType
+from snflwrpy.sorter_factory import sorter_factory
+from snflwrpy.sorters.bogo_sorter import BogoSorter
+from snflwrpy.sorters.bubble_sorter import BubbleSorter
+from snflwrpy.sorters.cocktail_sorter import CocktailSorter
+from snflwrpy.sorters.gnome_sorter import GnomeSorter
+from snflwrpy.sorters.insertion_sorter import InsertionSorter
+from snflwrpy.sorters.radix_lsd_sorter import RadixLSDSorter
+from snflwrpy.sort_type import SortType
+
+def test_sorter_factory_returns_bogo_sorter_when_requested():
+    """Test sorter factory returns bogo sorter when requested."""
+
+    sorter_name = SortType.BOGO.value
+    received_sorter = sorter_factory(sorter_name, [])
+    assert isinstance(received_sorter, BogoSorter)
 
 def test_sorter_factory_returns_bubble_sorter_when_requested():
     """Test sorter factory returns bubble sorter when requested."""