+++ vogel_sorter/__init__.py
@@ -1,3 +1,0 @@
-__name__ = 'Vogel Sorter'
-__version__ = '0.2.0'
-__description__ = "Visualize sorting algorithms with Vogel's Fibonacci model."